requirements.txt
requests==2.22.0
autopep8==1.5.2
Django==2.2.8
django-cors-headers==3.2.1
django-environ==0.4.5
django-generate-secret-key==1.0.2
django-templated-mail==1.1.1
django-webpack-loader==0.7.0
djangorestframework==3.10.3
djangorestframework-simplejwt==4.3.0
djoser==2.0.3
pycodestyle==2.6.0
PyJWT==1.7.1
python-dotenv==0.13.0
pytz==2020.1
sqlparse==0.3.1
termcolor==1.1.0
Pillow==6.2.0